Add files via upload
I added the loading, the spliting of the data in the main. The training and the testing seem to have problem because of something inside the Net (seems to be a dimension problem)

the net was also modified. the structure remains the same, but some things were without arguments or different "naming".